pyproject.toml
[tool.poetry]
name = "backendcourse"
version = "0.1.0"
description = "training course FastAPI"
authors = ["psk-a <ai3501669@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.35"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
greenlet = "^3.1.1"
pydantic-settings = "^2.5.2"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.9.2"}
redis = "^5.1.1"
fastapi-cache2 = "^0.2.2"
celery = "^5.4.0"
pillow = "^11.0.0"
python-multipart = "^0.0.17"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-dotenv = "^0.5.2"
httpx = "^0.27.2"
ruff = "^0.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
